+++ cmd/kube-apiserver/app/options/options_test.go
@@ -103,7 +103,8 @@ func TestAddFlags(t *testing.T) {
 			MinRequestTimeout:           1800,
 		},
 		Admission: &apiserveroptions.AdmissionOptions{
-			RecommendedPluginOrder: []string{"NamespaceLifecycle"},
+			RecommendedPluginOrder: []string{"NamespaceLifecycle", "Initializers"},
+			DefaultOffPlugins:      []string{"Initializers"},
 			PluginNames:            []string{"AlwaysDeny"},
 			ConfigFile:             "/admission-control-config",
 			Plugins:                s.Admission.Plugins,
