+++ staging/src/k8s.io/apiserver/pkg/server/options/admission_test.go
@@ -56,7 +56,7 @@ func TestEnabledPluginNamesMethod(t *testing.T) {
 			actualPluginNames := target.enabledPluginNames()
 
 			if len(actualPluginNames) != len(scenario.expectedPluginNames) {
-				t.Errorf("incorrect number of items, got %d, expected = %d", len(actualPluginNames), len(scenario.expectedPluginNames))
+				t.Fatalf("incorrect number of items, got %d, expected = %d", len(actualPluginNames), len(scenario.expectedPluginNames))
 			}
 			for i := range actualPluginNames {
 				if scenario.expectedPluginNames[i] != actualPluginNames[i] {
